[WIP] Fixed Infinite Loop Bug
This fixes the infinite loop found in the bitcoin_miner circuit.

This changes the backward and total cost calculations from float to
double. This also changes the heap such that it stores double instead of
float. The heap may be able to be turned into float; but, it requires
some unsavory code.

The total cost may also be able to remain a float; however, need to see
if master would change this to a double.

Also left in some debug code which prevents infinite loops so we do not
crash the testing machines. This can be removed once everything works.

+++ vpr/src/base/vpr_types.h
@@ -1734,8 +1734,8 @@ struct t_rr_node_route_inf {
     RREdgeId prev_edge;
 
     float acc_cost;
-    float path_cost;
-    float backward_path_cost;
+    double path_cost;
+    double backward_path_cost;
     float R_upstream; // TODO: Investigate the effect of adding the R_upstream field in
                       //       this struct. It is put in for the fine-grained parallel
                       //       router's benefits. It is increasing the working set, which
